+++ src/main/java/com/kodilla/ecommercee/dto/TestDto.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.dto;
+
+public class TestDto {
+}
