+++ src/main/java/com/kodilla/ecommercee/entity/TestEntity.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.entity;
+
+public class TestEntity {
+}
