+++ src/main/java/com/kodilla/ecommercee/exception/TestException.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.exception;
+
+public class TestException {
+}
