+++ src/main/java/com/kodilla/ecommercee/mapper/TestMapper.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.mapper;
+
+public class TestMapper {
+}
