+++ src/main/java/com/kodilla/ecommercee/repository/TestRepository.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.repository;
+
+public class TestRepository {
+}
