+++ src/main/java/com/kodilla/ecommercee/service/TestService.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.service;
+
+public class TestService {
+}
